Refactor some `devtools_shared` code
