+++ packages/devtools_app/lib/src/shared/primitives/utils.dart
@@ -1246,20 +1246,6 @@ Iterable<T> removeNullValues<T>(Iterable<T?> values) {
 /// the directory separators '/', and returning the last element.
 String? fileNameFromUri(String? uri) => uri?.split('/').last;
 
-String packageRootFromFileUriString(String fileUriString) {
-  // TODO(kenz): for robustness, consider sending the root library uri to the
-  // server and having the server look for the package folder that contains the
-  // `.dart_tool` directory.
-  final directoryRegExp =
-      RegExp(r'\/(lib|bin|integration_test|test|benchmark)\/.+\.dart');
-  final directoryIndex = fileUriString.indexOf(directoryRegExp);
-  if (directoryIndex != -1) {
-    fileUriString = fileUriString.substring(0, directoryIndex);
-  }
-  _log.fine('calculated rootFromFileUriString: $fileUriString');
-  return fileUriString;
-}
-
 /// Calculates subtraction of two maps.
 ///
 /// Result map keys is union of the imput maps' keys.
