+++ packages/devtools_app_shared/lib/src/ui/flex_split_column.dart
@@ -8,17 +8,17 @@ import 'split.dart';
 
 /// A widget that takes a list of [children] and lays them out in a column where
 /// each child has a flexible height.
-/// 
+///
 /// Each child in [children] must have an accompanying header in [header]. Since
 /// each header must be a [PreferredSizeWidget], it is common to use the shared
 /// widgets [AreaPaneHeader] or [BlankHeader] for the column [headers].
-/// 
+///
 /// The user can customize the amount of space allocated to each child by
 /// clicking and dragging the [header]s that separate the children.
 ///
 /// [initialFractions] defines how much space to give each child when building
 /// this widget.
-/// 
+///
 /// [minSizes] defines the minimum size that each child can be set to when
 /// adjusting the sizes of the children.
 final class FlexSplitColumn extends StatelessWidget {
