+++ packages/devtools_app_shared/lib/src/ui/split.dart
@@ -17,7 +17,7 @@ import '../utils/utils.dart';
 ///
 /// [initialFractions] defines how much space to give each child when building
 /// this widget.
-/// 
+///
 /// [minSizes] defines the minimum size that each child can be set to when
 /// adjusting the sizes of the children.
 final class Split extends StatefulWidget {
