+++ packages/devtools_shared/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 7.0.1-wip
 * Refactor yaml extension methods.
+* Add a helper method `packageRootFromFileUriString`.
 
 # 7.0.0
 * **Breaking change:** remove the `ServerApi.setCompleted` method that was a
