+++ packages/devtools_shared/lib/devtools_shared.dart
@@ -12,4 +12,5 @@ export 'src/memory/memory_json.dart';
 export 'src/service_utils.dart';
 export 'src/sse/sse_shim.dart';
 export 'src/utils/compare.dart';
+export 'src/utils/file_utils.dart';
 export 'src/utils/semantic_version.dart';
